(Storybook) Incorporated HSON Into SET
Like with the Unleashed Wii stuff, there's a lot of assumptions on the data here which is probably incorrect and the code isn't the tidiest, but it's something.

+++ KnuxLib/Helpers.cs
@@ -272,7 +272,7 @@ namespace KnuxLib
         /// Converts a floating point value to a signed integer stored in the Binary Angle Measurement System.
         /// </summary>
         /// <param name="value">The value to convert.</param>
-        public static float CalculateBAMsValue(float value) => value * 65536f / 360f;
+        public static int CalculateBAMsValue(float value) => (int)(value * 65536f / 360f);
     
         /// <summary>
         /// Creates a project for the HSON Specification in libHSON-csharp.
